@@ -1,4 +1,4 @@
-class CreatePages < ActiveRecord::Migration
+class CreatePages < ActiveRecord::Migration[5.0]
   def change
     create_table :pages do |t|
       t.string :title
