@@ -1,4 +1,4 @@
-class CreateInquiries < ActiveRecord::Migration
+class CreateInquiries < ActiveRecord::Migration[5.0]
   def change
     create_table :inquiries do |t|
       t.string 	:name
