@@ -1,4 +1,4 @@
-class ChangeCommentsStructure < ActiveRecord::Migration
+class ChangeCommentsStructure < ActiveRecord::Migration[5.0]
   def change
     remove_column :comments, :commenter, :string
     remove_column :comments, :body, :string
