@@ -1,4 +1,4 @@
-class CreateResponses < ActiveRecord::Migration
+class CreateResponses < ActiveRecord::Migration[5.0]
   def change
     create_table :responses do |t|
       t.belongs_to :comment
