@@ -1,4 +1,4 @@
-class AddSlugToPosts < ActiveRecord::Migration
+class AddSlugToPosts < ActiveRecord::Migration[5.0]
   def change
     add_column :posts, :slug, :string
     add_index :posts, :slug, unique: true
